lazy load telemetry entrypoint only when needed
