+++ packages/dd-trace/test/telemetry/index.spec.js
@@ -14,6 +14,73 @@ const DEFAULT_HEARTBEAT_INTERVAL = 60000
 
 let traceAgent
 
+describe('telemetry (proxy)', () => {
+  let telemetry
+  let proxy
+
+  beforeEach(() => {
+    telemetry = sinon.spy({
+      start () {},
+      stop () {},
+      updateIntegrations () {},
+      updateConfig () {},
+      appClosing () {}
+    })
+
+    proxy = proxyquire('../../src/telemetry', {
+      './telemetry': telemetry
+    })
+  })
+
+  it('should be noop when disabled', () => {
+    proxy.start()
+    proxy.updateIntegrations()
+    proxy.updateConfig()
+    proxy.appClosing()
+    proxy.stop()
+
+    expect(telemetry.start).to.not.have.been.called
+    expect(telemetry.updateIntegrations).to.not.have.been.called
+    expect(telemetry.updateConfig).to.not.have.been.called
+    expect(telemetry.appClosing).to.not.have.been.called
+    expect(telemetry.stop).to.not.have.been.called
+  })
+
+  it('should proxy when enabled', () => {
+    const config = { telemetry: { enabled: true } }
+
+    proxy.start(config)
+    proxy.updateIntegrations()
+    proxy.updateConfig()
+    proxy.appClosing()
+    proxy.stop()
+
+    expect(telemetry.start).to.have.been.calledWith(config)
+    expect(telemetry.updateIntegrations).to.have.been.called
+    expect(telemetry.updateConfig).to.have.been.called
+    expect(telemetry.appClosing).to.have.been.called
+    expect(telemetry.stop).to.have.been.called
+  })
+
+  it('should be noop when disabled after being enabled', () => {
+    const config = { telemetry: { enabled: true } }
+
+    proxy.start(config)
+    proxy.stop()
+    proxy.start()
+    proxy.updateIntegrations()
+    proxy.updateConfig()
+    proxy.appClosing()
+    proxy.stop()
+
+    expect(telemetry.start).to.have.been.calledOnce
+    expect(telemetry.updateIntegrations).to.not.have.been.called
+    expect(telemetry.updateConfig).to.not.have.been.called
+    expect(telemetry.appClosing).to.not.have.been.called
+    expect(telemetry.stop).to.have.been.calledOnce
+  })
+})
+
 describe('telemetry', () => {
   let telemetry
   let pluginsByName
@@ -39,7 +106,7 @@ describe('telemetry', () => {
 
     traceAgent.reqs = []
 
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -171,7 +238,7 @@ describe('telemetry', () => {
 
   it('should not send app-closing if telemetry is not enabled', () => {
     const sendDataStub = sinon.stub()
-    const notEnabledTelemetry = proxyquire('../../src/telemetry', {
+    const notEnabledTelemetry = proxyquire('../../src/telemetry/telemetry', {
       './send-data': {
         sendData: sendDataStub
       }
@@ -213,7 +280,7 @@ describe('telemetry app-heartbeat', () => {
         }
       }
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -278,7 +345,7 @@ describe('Telemetry extended heartbeat', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -320,7 +387,7 @@ describe('Telemetry extended heartbeat', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -463,7 +530,7 @@ describe('Telemetry retry', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -549,7 +616,7 @@ describe('Telemetry retry', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -618,7 +685,7 @@ describe('Telemetry retry', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -678,7 +745,7 @@ describe('Telemetry retry', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -760,7 +827,7 @@ describe('Telemetry retry', () => {
       }
 
     }
-    telemetry = proxyquire('../../src/telemetry', {
+    telemetry = proxyquire('../../src/telemetry/telemetry', {
       '../exporters/common/docker': {
         id () {
           return 'test docker id'
@@ -848,8 +915,8 @@ describe('AVM OSS', () => {
           traceAgent.reqs = []
 
           delete require.cache[require.resolve('../../src/telemetry/send-data')]
-          delete require.cache[require.resolve('../../src/telemetry')]
-          telemetry = require('../../src/telemetry')
+          delete require.cache[require.resolve('../../src/telemetry/telemetry')]
+          telemetry = require('../../src/telemetry/telemetry')
 
           telemetryConfig = {
             telemetry: { enabled: true, heartbeatInterval: HEARTBEAT_INTERVAL },
@@ -909,7 +976,7 @@ describe('AVM OSS', () => {
     }
 
     before(() => {
-      telemetry = proxyquire('../../src/telemetry', {
+      telemetry = proxyquire('../../src/telemetry/telemetry', {
         '../log': logSpy
       })
     })
